+++ src/components/launch-pad.js
@@ -20,7 +20,7 @@ import {
 import { useSpaceX } from "../utils/use-space-x";
 import Error from "./error";
 import Breadcrumbs from "./breadcrumbs";
-import { LaunchItem } from "./launches";
+import LaunchItem from "containers/LaunchesListPage/LaunchItem";
 
 export default function LaunchPad() {
   let { launchPadId } = useParams();
